requirements.txt
dj-database-url
Django
gunicorn
psycopg2
whitenoise
python-dotenv
django-ckeditor
misaka
django-resized
pillow
django-braces
numpy
django-anymail
django-recaptcha2
django-taggit
python-decouple